% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrvs.R
\name{lrvs}
\alias{lrvs}
\title{Perform linear regression variable selection}
\usage{
lrvs(
  X,
  y,
  sigma2,
  methods = c("ssvs", "bbssl", "debiased_lasso"),
  alpha = 0.05,
  verbose = T,
  ssvs_control = list(),
  bbssl_control = list(),
  debiased_lasso_control = list()
)
}
\arguments{
\item{X, y}{data matrix and response}

\item{sigma2}{known random error variance}

\item{methods}{any of: ssvs, bbssl, and debiased_lasso}

\item{alpha}{significance level, default 0.05}

\item{control}{control parameters for ssvs, bbssl and debiased_lasso. See ?ssvs, ?BB_SSL, ?SSLasso for control parameters}
}
\value{
a named list of methods. For ssvs, and bbsl: a list of posteriorMean and posterior credible set(a matrix, first column is lower ci, second one is upper ci.) For debiased_lasso: a list of beta_hat ,and confidence_interval.
}
\description{
Perform linear regression variable selection
}
