% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSLasso.R
\name{SSLasso}
\alias{SSLasso}
\title{debiased lasso}
\usage{
SSLasso(
  X,
  y,
  alpha = 0.05,
  lambda = NULL,
  mu = NULL,
  intercept = TRUE,
  resol = 1.3,
  maxiter = 50,
  threshold = 0.01,
  verbose = TRUE
)
}
\description{
debiased lasso
}
\details{
Compute confidence intervals and p-values.

Args:
  X     :  design matrix
  y     :  response
  alpha :  significance level
  lambda:  Lasso regularization parameter (if null, fixed by sqrt lasso)
  mu    :  Linfty constraint on M (if null, searches)
  resol :  step parameter for the function that computes M
  maxiter: iteration parameter for computing M
  threshold : tolerance criterion for computing M
  verbose : verbose?

Returns:
  noise.sd: Estimate of the noise standard deviation
  norm0   : Estimate of the number of 'significant' coefficients
  coef    : Lasso estimated coefficients
  unb.coef: Unbiased coefficient estimates
  low.lim : Lower limits of confidence intervals
  up.lim  : upper limit of confidence intervals
  pvals   : p-values for the coefficients
}
